requirements.txt
gradio==4.44.1
opencv-python==4.8.0.76
numpy==1.23.5
setuptools==58.2.0
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
mediapipe==0.10.11
matplotlib==3.5.1
scikit-learn==1.0.2
transformers==4.46.3
imageio==2.35.1